Seify Development (#33)
The primary work done in this PR is the splitup of the tx and rx sync streamers such that they can be generic for when the bladerf struct is wrapped with some sort of smart pointer. This is desired since it will likely need to be wrapped in an Arc for use with seify.